requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
faiss-cpu==1.7.4
filelock==3.24.3
Flask==2.3.3
fsspec==2026.2.0
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==1.24.3
packaging==26.0
pillow==12.1.1
PyPDF2==3.0.1
python-docx==0.8.11
PyYAML==6.0.3
regex==2026.2.19
requests==2.31.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.15.3
sentence-transformers==5.2.3
sentencepiece==0.2.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
